Back Handler success however broken on foward click
The idea of pushing the current page unto a stack and storing the URL in the navbar was implemented and would pop each page history from the stack when the back button was clicked. However, in some cases, the pages would load in their raw form when the forward button was clicked/the logo home button was clicked.